add email and team name validation
